refactored out flask-sqlalchemy calls
